refactor exchanges parser module
