+++ pyhodl/data/parsers.py
@@ -20,8 +20,9 @@
 
 from hal.files.models.system import ls_recurse, is_file
 
-from .core import CryptoParser, BinanceParser, BitfinexParser, CoinbaseParser, \
-    GdaxParser
+from pyhodl.data.exchanges import BinanceParser, BitfinexParser, \
+    CoinbaseParser, GdaxParser
+from .core import CryptoParser
 
 
 def build_parser(input_file):
